Rename libmastermind to mastermind-cache
This have been long coming -- name like libENTITY imply that this is
either implementation of ENTITY or client of ENTITY.

But `libmastermind` is neither.

Then what it really is?

It's an implementation of in-process cache for mastermind data providing:
- C++ interface for C++ clients and
- Python binding for python clients on top of C++ interface

So proper name for it is `mastermind_cache`.

(Change of name involves packages, include directory, dynamic libraries etc.)

MDS-2573: Наконец переименовать libmastermind в mastermind-cache